agent/requirements.txt
anthropic==0.66.0
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.75
langchain_google_genai==2.1.10
langchain_huggingface==0.3.1
langchain_openai==0.3.32
neo4j==5.28.1
openai==1.106.1
python-dotenv==1.1.1
Requests==2.32.5
